requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
redis==5.0.1
aioboto3==12.0.0
python-chess==1.999
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
websockets==12.0